Fix imlicit conversion warning SYCL

+++ core/src/SYCL/Kokkos_SYCL_TeamPolicy.hpp
@@ -116,7 +116,8 @@ class Kokkos::Impl::TeamPolicyInternal<Kokkos::SYCL, Properties...>
     const size_t max_reserved_shared_mem_per_team =
         (max_possible_team_size + 2) * sizeof(double);
     // arbitrarily setting level 1 scratch limit to 20MB
-    constexpr size_t max_l1_scratch_size = 20 * 1024 * 1024;
+    constexpr size_t max_l1_scratch_size =
+        static_cast<size_t>(20) * 1024 * 1024;
 
     size_t max_shmem = sycl_instance.m_maxShmemPerBlock;
     return (level == 0 ? max_shmem - max_reserved_shared_mem_per_team
